+++ src/Modules/Guilds/Wrappers/GuildSidebarWrapper.tsx
@@ -1,12 +1,12 @@
 import { useTypedParams } from 'Modules/Guilds/Hooks/useTypedParams';
-import { useGuildConfig } from 'hooks/Guilds/ether-swr/guild/useGuildConfig';
-import useGuildMemberTotal from 'hooks/Guilds/ether-swr/guild/useGuildMemberTotal';
-import { useVotingPowerOf } from 'hooks/Guilds/ether-swr/guild/useVotingPowerOf';
+import { useGuildConfig } from 'hooks/Guilds/guild/useGuildConfig';
+import useGuildMemberTotal from 'hooks/Guilds/guild/useGuildMemberTotal';
+import { useVotingPowerOf } from 'hooks/Guilds/guild/useVotingPowerOf';
 import { GuildSidebar } from 'components/GuildSidebar';
 import { MemberActions } from 'components/GuildSidebar/MemberActions';
 import { GuestActions } from 'components/GuildSidebar/GuestActions';
 import { useERC20Info } from 'hooks/Guilds/ether-swr/erc20/useERC20Info';
-import { useVoterLockTimestamp } from 'hooks/Guilds/ether-swr/guild/useVoterLockTimestamp';
+import { useVoterLockTimestamp } from 'hooks/Guilds/guild/useVoterLockTimestamp';
 import useGuildImplementationType from 'hooks/Guilds/guild/useGuildImplementationType';
 import { useTransactions } from 'contexts/Guilds';
 import { useERC20Guild } from 'hooks/Guilds/contracts/useContract';
@@ -14,7 +14,7 @@ import { formatUnits } from 'ethers/lib/utils';
 import useVotingPowerPercent from 'hooks/Guilds/guild/useVotingPowerPercent';
 import { useState } from 'react';
 import { WalletModal } from 'components/Web3Modals';
-import useTotalLocked from 'hooks/Guilds/ether-swr/guild/useTotalLocked';
+import useTotalLocked from 'hooks/Guilds/guild/useTotalLocked';
 import StakeTokensModalWrapper from './StakeTokensModalWrapper';
 import { useAccount, useEnsAvatar, useEnsName } from 'wagmi';
 import { isReadOnly } from 'provider/wallets';
