+++ src/Modules/Guilds/pages/LandingPage/LandingPage.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { useGuildRegistry } from 'hooks/Guilds/ether-swr/guild/useGuildRegistry';
+import { useGuildRegistry } from 'hooks/Guilds/guild/useGuildRegistry';
 import { GuildCard } from 'components/GuildCard/GuildCard';
 
-import useGuildMemberTotal from 'hooks/Guilds/ether-swr/guild/useGuildMemberTotal';
-import useActiveProposalsNow from 'hooks/Guilds/ether-swr/guild/useGuildActiveProposals';
+import useGuildMemberTotal from 'hooks/Guilds/guild/useGuildMemberTotal';
+import useActiveProposalsNow from 'hooks/Guilds/guild/useGuildActiveProposals';
 import useENSNameFromAddress from 'hooks/Guilds/ens/useENSNameFromAddress';
-import { useGuildConfig } from 'hooks/Guilds/ether-swr/guild/useGuildConfig';
+import { useGuildConfig } from 'hooks/Guilds/guild/useGuildConfig';
 
 import { CardsContainer } from './LandingPage.styled';
