+++ src/components/ProposalInfoCard/types.tsx
@@ -1,5 +1,5 @@
 import { Proposal } from 'types/types.guilds.d';
-import { GuildConfigProps } from 'hooks/Guilds/ether-swr/guild/useGuildConfig';
+import { GuildConfigProps } from 'hooks/Guilds/guild/useGuildConfig';
 
 export interface ProposalInfoCardProps {
   proposal: Proposal;
