+++ src/components/ProposalVoteCard/types.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from 'ethers';
-import type { VoteData } from 'hooks/Guilds/ether-swr/guild/useVotingResults';
+import type { VoteData } from 'hooks/Guilds/guild/useVotingResults';
 import { ERC20Guild } from 'types/contracts';
 import { ProposalMetadata } from 'types/types.guilds';
 import { Moment } from 'moment';
