+++ src/hooks/Guilds/guild/useProposalCalls.ts
@@ -3,8 +3,8 @@ import { useTheme } from 'styled-components';
 import { useTranslation } from 'react-i18next';
 import { bulkDecodeCallsFromOptions } from '../contracts/useDecodedCall';
 import { decodeCall } from 'hooks/Guilds/contracts/useDecodedCall';
-import { useProposal } from '../ether-swr/guild/useProposal';
-import { useVotingResults } from 'hooks/Guilds/ether-swr/guild/useVotingResults';
+import { useProposal } from './useProposal';
+import { useVotingResults } from 'hooks/Guilds/guild/useVotingResults';
 import { Call, Option } from 'components/ActionsBuilder/types';
 import { ZERO_HASH } from 'utils';
 import useProposalMetadata from '../useProposalMetadata';
