+++ src/hooks/Guilds/guild/useTotalLocked.ts
@@ -1,12 +1,12 @@
 import { BigNumber } from 'ethers';
-import useEtherSWR from '../useEtherSWR';
+import useEtherSWR from '../ether-swr/useEtherSWR';
 import useCurrentSnapshotId from './useCurrentSnapshotId';
 import useGuildToken from './useGuildToken';
 import useTotalSupplyAt from './useTotalSupplyAt';
 import { useTypedParams } from 'Modules/Guilds/Hooks/useTypedParams';
 import ERC20GuildContract from 'contracts/ERC20Guild.json';
-import useSnapshotId from 'hooks/Guilds/ether-swr/guild/useSnapshotId';
-import useTotalLockedAt from 'hooks/Guilds/ether-swr/guild/useTotalLockedAt';
+import useSnapshotId from 'hooks/Guilds/guild/useSnapshotId';
+import useTotalLockedAt from 'hooks/Guilds/guild/useTotalLockedAt';
 import useGuildImplementationType from 'hooks/Guilds/guild/useGuildImplementationType';
 
 const useTotalLocked = (guildAddress: string, snapshotId?: string) => {
