+++ src/hooks/Guilds/guild/useVotingPowerOf.ts
@@ -1,7 +1,7 @@
 import { BigNumber } from 'ethers';
 import { SWRResponse } from 'swr';
-import useEtherSWR from '../useEtherSWR';
-import useGuildImplementationType from '../../guild/useGuildImplementationType';
+import useEtherSWR from '../ether-swr/useEtherSWR';
+import useGuildImplementationType from './useGuildImplementationType';
 import useVotingPowerOfAt from './useVotingPowerOfAt';
 import ERC20GuildContract from 'contracts/ERC20Guild.json';
