+++ src/utils/fixtures.ts
@@ -3,7 +3,7 @@ import { BigNumber } from 'ethers';
 import moment from 'moment';
 import { Proposal, ContractState, ProposalState } from 'types/types.guilds.d';
 import { ENSAvatar } from 'types/types.guilds';
-import { GuildConfigProps } from 'hooks/Guilds/ether-swr/guild/useGuildConfig';
+import { GuildConfigProps } from 'hooks/Guilds/guild/useGuildConfig';
 import {
   Call,
   DecodedAction,
